refactor: migrate common/components to src/components with renamed form components
