+++ src/components/FormPasswordInput.tsx
@@ -1,10 +1,12 @@
 import { FieldError, Path, UseFormRegister, ValidationRule } from 'react-hook-form';
 import { IFormValues } from 'app/core/types';
-import './TextInput.scss';
+
+import './FormPasswordInput.scss';
+import { Eye, EyeSlash } from '@phosphor-icons/react';
+import { useState } from 'react';
 
 interface InputProps {
   label: Path<IFormValues>;
-  type: 'text' | 'email' | 'number';
   disabled?: boolean;
   register: UseFormRegister<IFormValues>;
   minLength?: ValidationRule<number>;
@@ -14,17 +16,17 @@ interface InputProps {
   error?: FieldError;
   min?: ValidationRule<number | string>;
   required?: boolean;
-  className?: string;
-  autoFocus?: boolean;
   onFocus?: () => void;
   onBlur?: () => void;
+  className?: string;
+  autoFocus?: boolean;
+  value?: string;
   autoComplete?: string;
   inputDataCy?: string;
 }
 
-export default function TextInput({
+const PasswordInput = ({
   label,
-  type,
   disabled,
   register,
   minLength,
@@ -34,22 +36,25 @@ export default function TextInput({
   error,
   min,
   required,
+  onFocus,
+  onBlur,
   className,
   autoFocus,
   autoComplete,
   inputDataCy,
-}: Readonly<InputProps>): JSX.Element {
+}: InputProps): JSX.Element => {
+  const [showPassword, setShowPassword] = useState(false);
+
   return (
-    <div className={`${className}`}>
+    <div className={`relative flex-1 ${className}`}>
       <input
-        type={type}
+        type={showPassword ? 'text' : 'password'}
         disabled={disabled}
         placeholder={placeholder}
-        autoComplete={autoComplete}
-        id={label}
         min={0}
         required={true}
         autoFocus={autoFocus}
+        autoComplete={autoComplete}
         data-cy={inputDataCy}
         {...register(label, {
           required,
@@ -58,8 +63,23 @@ export default function TextInput({
           maxLength,
           pattern,
         })}
+        onFocus={() => {
+          if (onFocus) onFocus();
+        }}
+        onBlur={() => {
+          if (onBlur) onBlur();
+        }}
         className={error ? 'inxt-input input-error' : 'inxt-input input-primary'}
       />
+      <button
+        type="button"
+        onClick={() => setShowPassword(!showPassword)}
+        className="absolute right-4 top-1/2 flex -translate-y-1/2 cursor-pointer items-center justify-center text-gray-100"
+      >
+        {showPassword ? <Eye className="h-6 w-6" /> : <EyeSlash className="h-6 w-6" />}
+      </button>
     </div>
   );
-}
+};
+
+export default PasswordInput;
