+++ src/views/Login/components/LogIn.tsx
@@ -25,8 +25,8 @@ import workspacesService from 'app/core/services/workspace.service';
 import notificationsService, { ToastType } from 'app/notifications/services/notifications.service';
 import useLoginRedirections from '../hooks/useLoginRedirections';
 import shareService from 'app/share/services/share.service';
-import PasswordInput from 'common/components/PasswordInput';
-import TextInput from 'common/components/TextInput';
+import FormPasswordInput from 'components/FormPasswordInput';
+import FormTextInput from 'components/FormTextInput';
 import { AuthMethodTypes } from 'views/Checkout/types';
 import vpnAuthService from 'services/vpnAuth.service';
 import envService from 'app/core/services/env.service';
@@ -229,7 +229,7 @@ export default function LogIn(): JSX.Element {
         </h1>
 
         <form data-cy="loginWrapper" className="flex w-full flex-col space-y-2" onSubmit={handleSubmit(onSubmit)}>
-          <TextInput
+          <FormTextInput
             placeholder={translate('auth.email')}
             inputDataCy="emailInput"
             label="email"
@@ -239,7 +239,7 @@ export default function LogIn(): JSX.Element {
             error={errors.email}
           />
 
-          <PasswordInput
+          <FormPasswordInput
             placeholder={translate('auth.password')}
             inputDataCy="passwordInput"
             label="password"
@@ -250,7 +250,7 @@ export default function LogIn(): JSX.Element {
           />
 
           {showTwoFactor && (
-            <PasswordInput
+            <FormPasswordInput
               label="twoFactorCode"
               placeholder={translate('auth.login.twoFactorAuthenticationCode')}
               error={errors.twoFactorCode}
