+++ src/views/Signup/WorkspaceGuestSignUpView.test.tsx
@@ -93,7 +93,7 @@ describe('onSubmit', () => {
       X: () => <div>Mocked X Icon</div>,
     }));
 
-    vi.mock('common/components/PasswordInput', () => {
+    vi.mock('components/FormPasswordInput', () => {
       return {
         __esModule: true,
         default: vi.fn(({ register, ...props }) => (
