+++ vitest.config.ts
@@ -7,7 +7,7 @@ export default defineConfig({
       app: path.resolve(__dirname, './src/app'),
       views: path.resolve(__dirname, './src/views'),
       services: path.resolve(__dirname, './src/services'),
-      common: path.resolve(__dirname, './src/common'),
+      components: path.resolve(__dirname, './src/components'),
       crypto: 'crypto-browserify',
       stream: 'stream-browserify',
       path: 'path-browserify',
